requirements.txt
django==4.0.2
factory-boy==3.2.1
